Changed Field Type for Prices

+++ apps/products/models.py
@@ -38,7 +38,7 @@ class ProductManager(models.Manager):
 class Product(models.Model):
     photo = models.CharField(max_length=1000)
     product_name = models.CharField(max_length=300)
-    product_starting_bid = models.IntegerField(max_length=300)
+    product_starting_bid = models.FloatField(max_length=300)
     product_description = models.CharField(max_length=500)
     product_end_date = models.DateTimeField()
     owner = models.ForeignKey(User)
